requirements.txt
datasets==3.1.0
numpy==1.24
fire==0.7.0
huggingface-hub==0.25.0
jieba==0.42.1
Jinja2==3.1.4
matplotlib==3.9.2
ninja==1.11.1.1
nltk==3.9.1
tiktoken==0.7.0
tokenizers==0.20.3
tqdm==4.67.0
transformers==4.46.1
transformers-stream-generator==0.0.5
InstructorEmbedding
sentence-transformers==2.2.2
torch==2.5.0
faiss-gpu
protobuf
rouge